Changed from decimal to int for maxLoanMultiplier

+++ BankApp/SystemOwner.cs
@@ -10,7 +10,7 @@ namespace BankApp
     {
         public string OwnerID { get; set; } // Unique identifier for the system owner
         public string Name { get; set; } // System owner's full name
-        public decimal MaxLoanMultiplier { get; set; } = 5; // e.g., 5 means max loan = 5 × total deposits
+        public int MaxLoanMultiplier { get; set; } = 5; // e.g., 5 means max loan = 5 × total deposits
 
 
         // List<Account> is the collection of all accounts to loop through
@@ -33,7 +33,7 @@ namespace BankApp
         }
 
         // e.g., “max loan = 5× total deposits”
-        public void SetLoanPolicy(decimal maxLoanMultiplier)
+        public void SetLoanPolicy(int maxLoanMultiplier)
         {
             MaxLoanMultiplier = maxLoanMultiplier;
         }
